Tomcat Servlet Container and a HelloWorld java app created out of the old Tomcat demo app
Tomcat App now contains servlet container only. It deploys Tomcat on the instance and configures appropriate security rules
(8080 port added). The logic to deploy particular applications on it was removed and placed to a different package.

HelloWorld java application is created, having references on both PostgreSQL and Tomcat, being able to deploy the app on top of the Tomcat, call the database server to create a DB and user and configure its config file.