Add abstract Tomcat cluster  application
This patch adds new  Tomcat cluster  application to implement Tomcat Servlet Container interface. This application creates a cluster of VM and installs a tomcat service on them. This app creates a Load balancer pool and puts created instances to this LB pool as pool members. The app returns LB floating IP as a public IP.